Android compiler does not like qMax(qreal, double) added explicit typecasts

+++ GraphicsItems/alternationgraphicsitem.cpp
@@ -62,7 +62,7 @@ QRectF AlternationGraphicsItem::boundingRect() const
     double horizontalPadding = settings.value("Visualisation/Alternation/HorizontalPadding", 8.0).toDouble();
     QSizeF childrenRect = _mainLayout->sizeHint(Qt::PreferredSize);
 
-    return QRectF(0, 0, qMax(childrenRect.width(), 2*horizontalPadding + _title->boundingRect().width()), childrenRect.height());
+    return QRectF(0, 0, qMax(static_cast<double>(childrenRect.width()), 2*horizontalPadding + _title->boundingRect().width()), childrenRect.height());
 }
 
 void AlternationGraphicsItem::paint(QPainter *painter, const QStyleOptionGraphicsItem *option, QWidget *widget)
